# perf: writeToBuffer