--- a/.eslintrc
+++ b/.eslintrc
@@ -6,6 +6,10 @@
     "max-len": [
       "error",
       { "code": 80 }
+    ],
+    "quotes": [
+      "error",
+      "single"
     ]
   },
   "overrides": [