--- a/.gitignore
+++ b/.gitignore
@@ -30,3 +30,4 @@ node_modules
 # Vim
 *.swp
 
+package-lock.json