--- a/lib/tags/byte.js
+++ b/lib/tags/byte.js
@@ -29,11 +29,9 @@ class TAGByte extends BaseTag {
    * Write body to buffer
    * @param {Buffer} buff The buffer to write to
    * @param {number} offset The offset to start writing to
-   * @return {number} The number of bytes written
    */
-  writeBuffer(buff, offset) {
-    buff.writeInt8(this.value, offset);
-    return 1;
+  writeBuffer(buff) {
+    buff.writeInt8(this.value);
   }
 
   /**