--- a/lib/tags/byte_array.js
+++ b/lib/tags/byte_array.js
@@ -12,6 +12,7 @@ class TAGByteArray extends BaseTag {
   constructor() {
     super();
     this.type = 'TAG_Byte_Array';
+    this.value = [];
   }
 
   /**
@@ -43,18 +44,15 @@ class TAGByteArray extends BaseTag {
 
   /**
    * Write body to buffer
-   * @param {Buffer} buff The buffer to write to
+   * @param {Bvffer} buff The buffer to write to
    * @param {number} offset The offset to start writing from
-   * @return {number} The number of bytes written
    */
-  writeBuffer(buff, offset) {
-    buff.writeUInt32BE(this.value.length, offset);
+  writeBuffer(buff) {
+    buff.writeUInt32BE(this.value.length);
 
     for (let i = 0; i < this.value.length; i++) {
-      buff.writeInt8(this.value[i], offset + 4 + i);
+      buff.writeInt8(this.value[i]);
     }
-
-    return 4 + this.value.length;
   }
 
   /**