--- a/lib/tags/compound.js
+++ b/lib/tags/compound.js
@@ -85,8 +85,7 @@ class TAGCompound extends BaseTag {
 
       // child type id for 1 byte, child name length for 2 bytes and child
       // name for (child name length) byte(s).
-      len += 1;
-      len += 2;
+      len += 3;
       len += Buffer.byteLength(this.value[key].id, 'utf8');
 
       // add the child body's length
@@ -101,29 +100,23 @@ class TAGCompound extends BaseTag {
 
   /**
    * Write body to buffer
-   * @param {Buffer} buff The buffer to write to
+   * @param {Bvffer} buff The buffer to write to
    * @param {number} offset The offset to start writing to
-   * @return {number} The number of bytes written
    */
-  writeBuffer(buff, offset) {
-    let len = 0;
+  writeBuffer(buff) {
     for (const key in this.value) {
       if (!this.value.hasOwnProperty(key)) continue;
 
       const object = this.value[key];
-      buff.writeUInt8(object.getTypeId(), offset + len);
+      buff.writeUInt8(object.getTypeId());
 
-      const nameBuff = new Buffer(object.id, 'utf8');
-      nameBuff.copy(buff, offset + len + 1 + 2);
-      buff.writeUInt16BE(nameBuff.length, offset + len + 1);
+      const nameBuff = Buffer.from(object.id, 'utf8');
+      buff.writeBuffer(nameBuff);
 
-      len += object.writeBuffer(buff, offset + len + 1 + 2 + nameBuff.length);
-      len += (1 + 2 + nameBuff.length);
+      object.writeBuffer(buff);
     }
 
-    buff.writeUInt8(0, offset + len);
-    len += 1;
-    return len;
+    buff.writeUInt8(0);
   }
 
   /**