--- a/lib/tags/double.js
+++ b/lib/tags/double.js
@@ -28,12 +28,9 @@ class TAGDouble extends BaseTag {
   /**
    * Write body to buffer
    * @param {Buffer} buff The buffer to write to
-   * @param {number} offset The offset to start writing to
-   * @return {number} The number of bytes written
    */
-  writeBuffer(buff, offset) {
-    buff.writeDoubleBE(this.value, offset);
-    return 8;
+  writeBuffer(buff) {
+    buff.writeDoubleBE(this.value);
   }
 
   /**