--- a/lib/tags/float.js
+++ b/lib/tags/float.js
@@ -28,12 +28,9 @@ class TAGFloat extends BaseTag {
   /**
    * Write body to buffer
    * @param {Buffer} buff The buffer to write to
-   * @param {number} offset The offset to start writing to
-   * @return {number} The number of bytes written
    */
-  writeBuffer(buff, offset) {
-    buff.writeFloatBE(this.value, offset);
-    return 4;
+  writeBuffer(buff) {
+    buff.writeFloatBE(this.value);
   }
 
   /**