--- a/lib/tags/int.js
+++ b/lib/tags/int.js
@@ -27,13 +27,10 @@ class TAGInt extends BaseTag {
 
   /**
    * Write body to buffer
-   * @param {Buffer} buff The buffer to write to
-   * @param {number} offset The offset to start writing to
-   * @return {number} The number of bytes written
+   * @param {Bvffer} buff The buffer to write to
    */
-  writeBuffer(buff, offset) {
-    buff.writeInt32BE(this.value, offset);
-    return 4;
+  writeBuffer(buff) {
+    buff.writeInt32BE(this.value);
   }
 
   /**