--- a/lib/tags/int_array.js
+++ b/lib/tags/int_array.js
@@ -136,17 +136,13 @@ class TAGIntArray extends BaseTag {
   /**
    * Write body to buffer
    * @param {Buffer} buff The buffer to write to
-   * @param {number} offset The offset to start writing from
-   * @return {number} The number of bytes written
    */
-  writeBuffer(buff, offset) {
-    buff.writeUInt32BE(this.value.length, offset);
+  writeBuffer(buff) {
+    buff.writeUInt32BE(this.value.length);
 
     for (let i = 0; i < this.value.length; i++) {
-      buff.writeInt32BE(this.value[i], offset + 4 + i * 4);
+      buff.writeInt32BE(this.value[i]);
     }
-
-    return 4 + this.value.length * 4;
   }
 }