--- a/lib/tags/list.js
+++ b/lib/tags/list.js
@@ -24,6 +24,7 @@ class TAGList extends BaseTag {
     this.type = 'TAG_List';
     this.childType = 'TAG_End';
     this.$tags = getTags();
+    this.value = [];
   }
 
   /**
@@ -53,10 +54,9 @@ class TAGList extends BaseTag {
     this.value = [];
     let nextOffset = offset + 1 + 4;
     for (let i = 0; i < len; i++) {
-      const element = (typeId === 0) ? undefined : new Tag();
-      const elementLength = (typeId === 0) ?
-        0 :
-        element._readBodyFromBuffer(buff, nextOffset);
+      const element = typeId === 0 ? undefined : new Tag();
+      const elementLength =
+        typeId === 0 ? 0 : element._readBodyFromBuffer(buff, nextOffset);
       nextOffset += elementLength;
 
       this.value.push(element);
@@ -72,7 +72,8 @@ class TAGList extends BaseTag {
   calcBufferLength() {
     return this.value.reduce(
       (sum, child) => sum + child.calcBufferLength(),
-      1 + 4);
+      1 + 4
+    );
   }
 
   /**
@@ -104,7 +105,8 @@ class TAGList extends BaseTag {
     for (let i = 0; i < value.length; i++) {
       if (!(value[i] instanceof TagType)) {
         throw new TypeError(
-          `Inconsistent TAG_List element type at position ${i}.`);
+          `Inconsistent TAG_List element type at position ${i}.`
+        );
       }
 
       array.push(value[i]);
@@ -220,26 +222,21 @@ class TAGList extends BaseTag {
   /**
    * Write body to buffer
    * @param {Buffer} buff The buffer to write to
-   * @param {number} offset The offset to start writing from
-   * @return {number} The number of bytes written
    */
-  writeBuffer(buff, offset) {
+  writeBuffer(buff) {
     // no element
     if (!this.value.length) {
-      buff.writeUInt8(0, offset);
-      buff.writeUInt32BE(0, offset + 1);
-      return 1 + 4;
+      buff.writeUInt8(0);
+      buff.writeUInt32BE(0);
+      return;
     }
 
-    buff.writeUInt8(this.value[0].getTypeId(), offset);
-    buff.writeUInt32BE(this.value.length, offset + 1);
-    let len = 0;
-    const baseOffset = offset + 1 + 4;
+    buff.writeUInt8(this.value[0].getTypeId());
+    buff.writeUInt32BE(this.value.length);
+
     for (let i = 0; i < this.value.length; i++) {
-      len += this.value[i].writeBuffer(buff, baseOffset + len);
+      this.value[i].writeBuffer(buff);
     }
-
-    return len + 1 + 4;
   }
 }