--- a/lib/tags/long.js
+++ b/lib/tags/long.js
@@ -20,12 +20,11 @@ class TAGLong extends BaseTag {
     return 8;
   }
 
-  writeBuffer(buff, offset) {
+  writeBuffer(buff) {
     const bytes = this.value.toBytesBE();
-    for (let i = 0, j = offset; i < 8; i++, j++) {
-      buff.writeUInt8(bytes[i], j);
+    for (let i = 0; i < 8; i++) {
+      buff.writeUInt8(bytes[i]);
     }
-    return 8;
   }
 
   toJSON() {