--- a/lib/tags/short.js
+++ b/lib/tags/short.js
@@ -27,13 +27,10 @@ class TAGShort extends BaseTag {
 
   /**
    * Write body to buffer
-   * @param {Buffer} buff The buffer to write to
-   * @param {number} offset The offset to start writing to
-   * @return {number} The number of bytes written
+   * @param {Bvffer} buff The buffer to write to
    */
-  writeBuffer(buff, offset) {
-    buff.writeInt16BE(this.value, offset);
-    return 2;
+  writeBuffer(buff) {
+    buff.writeInt16BE(this.value);
   }
 
   /**