--- a/lib/tags/string.js
+++ b/lib/tags/string.js
@@ -39,15 +39,11 @@ class TAGString extends BaseTag {
    * Write body to buffer
    * @param {Buffer} buff The buffer to write to
    * @param {number} offset The offset to start writing to
-   * @return {number} The number of bytes written
    */
-  writeBuffer(buff, offset) {
-    const strBuff = new Buffer(this.value, 'utf8');
+  writeBuffer(buff) {
+    const strBuff = Buffer.from(this.value, 'utf8');
 
-    strBuff.copy(buff, offset + 2);
-    buff.writeUInt16BE(strBuff.length, offset);
-
-    return 2 + strBuff.length;
+    buff.writeBuffer(strBuff);
   }
 
   /**