--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "mcnbt",
-  "version": "2.0.1",
+  "version": "2.0.2",
   "description": "Yet another Minecraft NBT format file / buffer parser for Node.js.",
   "main": "nbt.js",
   "scripts": {