v: 2.0.1 , r: 1b x 66,319 ops/sec ±70.40% (88 runs sampled)
v: 2.0.1 , r: 1kb x 32,589 ops/sec ±0.96% (90 runs sampled)
v: 2.0.1 , r: 2kb x 18,932 ops/sec ±0.51% (93 runs sampled)
v: 2.0.1 , r: 4kb x 10,554 ops/sec ±0.44% (94 runs sampled)
v: 2.0.1 , r: 8kb x 5,425 ops/sec ±0.56% (91 runs sampled)

v: 2.0.2 , r: 1b x 159,682 ops/sec ±0.49% (92 runs sampled)
v: 2.0.2 , r: 1kb x 35,753 ops/sec ±0.57% (90 runs sampled)
v: 2.0.2 , r: 2kb x 19,579 ops/sec ±0.67% (95 runs sampled)
v: 2.0.2 , r: 4kb x 10,397 ops/sec ±0.48% (95 runs sampled)
v: 2.0.2 , r: 8kb x 5,168 ops/sec ±0.35% (95 runs sampled)